introduce AbstractDatabaseOperationEvent
because the Pre and Post versions were isomorphic
