+++ hibernate-core/src/main/java/org/hibernate/event/spi/AbstractPreDatabaseOperationEvent.java
@@ -4,7 +4,6 @@
  */
 package org.hibernate.event.spi;
 
-import org.hibernate.engine.spi.SessionFactoryImplementor;
 import org.hibernate.persister.entity.EntityPersister;
 
 /**
@@ -12,11 +11,7 @@ import org.hibernate.persister.entity.EntityPersister;
  *
  * @author Steve Ebersole
  */
-public abstract class AbstractPreDatabaseOperationEvent extends AbstractEvent {
-
-	private final Object entity;
-	private final Object id;
-	private final EntityPersister persister;
+public abstract class AbstractPreDatabaseOperationEvent extends AbstractDatabaseOperationEvent {
 
 	/**
 	 * Constructs an event containing the pertinent information.
@@ -31,46 +26,6 @@ public abstract class AbstractPreDatabaseOperationEvent extends AbstractEvent {
 			Object entity,
 			Object id,
 			EntityPersister persister) {
-		super( source );
-		this.entity = entity;
-		this.id = id;
-		this.persister = persister;
-	}
-
-	/**
-	 * Retrieves the entity involved in the database operation.
-	 *
-	 * @return The entity.
-	 */
-	public Object getEntity() {
-		return entity;
-	}
-
-	/**
-	 * The id to be used in the database operation.
-	 *
-	 * @return The id.
-	 */
-	public Object getId() {
-		return id;
-	}
-
-	/**
-	 * The persister for the entity.
-	 *
-	 * @return The entity persister.
-	 */
-	public EntityPersister getPersister() {
-		return persister;
-	}
-
-	/**
-	 * The factory which owns the persister for the entity.
-	 *
-	 * @return The factory
-	 */
-	@Override
-	public SessionFactoryImplementor getFactory() {
-		return persister.getFactory();
+		super( source, entity, id, persister );
 	}
 }
